backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.7
aioredis==2.0.1

# AI / ML
transformers==4.41.2
torch==2.3.1
sentence-transformers==3.0.1
accelerate==0.31.0
huggingface-hub==0.23.4
numpy==1.26.4
scikit-learn==1.5.0

# Task Queue
celery==5.4.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
orjson==3.10.5
tenacity==8.4.1
